renamed and changed place of WorkoutViewModelNew to WorkoutViewModel and moved it to unused
